+++ src/app/Repositories/CommentRepository.php
@@ -3,12 +3,20 @@
 namespace App\Repositories;
 
 use App\Models\Comment;
+use Dingo\Api\Exception\ResourceException;
 
 class CommentRepository
 {
-    public function getComments($articleId, $userId)
+    public function getComments($articleId)
     {
-        return Comment::whereUserId($userId)->whereArticleId($articleId);
+        $comments = Comment::where('articleId', $articleId)->get();
+
+        // If the comment doesn't exist, throw an exception
+        if (!$comments  || count($comments) === 0) {
+            throw new ResourceException('Comment not found');
+        }
+
+        return $comments;
     }
 
     public function addComment($commentTitle, $commentContent, $userId, $articleId)
@@ -18,12 +26,23 @@ class CommentRepository
         $comment->content = $commentContent;
         $comment->userId = $userId;
         $comment->articleId = $articleId;
-        return $comment->save();
+
+        try {
+            return $comment->save();
+        } catch (\Exception $e) {
+            throw new ResourceException('Comment not added');
+        }
     }
 
-    public function deleteComment($commentId)
+    public function deleteComment($commentId, $articleId)
     {
         $comment = Comment::find($commentId);
+
+        // If the comment doesn't exist, throw an exception
+        if (!$comment) {
+            throw new ResourceException('Comment not found');
+        }
+
         return $comment->delete();
     }
 }
